Web: Files: refactoring: created withBadges HOC, rename withLoader, applied changes
